archive/requirements.txt
algoliasearch>=4,<5
beautifulsoup4
gpxpy
lxml
python-dotenv
pyproj
requests
tqdm
